[SKP-147] refac : useAnalyze 훅 분리 1
